refactor: move uriFactory to its own file
